Refactor the compiler, too
